+++ tasks/omp/vinicuk_t_complex_sparse_matrix_mult_csc/include/csc_complex_matrix.hpp
@@ -7,13 +7,13 @@
 
 struct CSCComplexMatrix {
   std::vector<std::complex<double>> values;
-  std::vector<size_t> row_indexes;
-  std::vector<size_t> col_ptrs;
+  std::vector<int> row_indexes;
+  std::vector<int> col_ptrs;
 
-  size_t num_rows;
-  size_t num_cols;
+  int num_rows;
+  int num_cols;
 
   // CSSComplexMatrix(size_t n_rows, size_t n_cols);
 
-  CSCComplexMatrix(size_t _num_cols, size_t _num_rows) : num_rows(_num_rows), num_cols(_num_cols) {}
+  CSCComplexMatrix(int _num_cols, int _num_rows) : num_rows(_num_rows), num_cols(_num_cols) {}
 };
